Increase map mask size and add mimic flag.

+++ crawl-ref/source/externs.h
@@ -82,7 +82,7 @@ const int kPathLen = 256;
 #define NO_BERSERK_PENALTY    -1
 
 typedef FixedArray<dungeon_feature_type, GXM, GYM> feature_grid;
-typedef FixedArray<unsigned short, GXM, GYM> map_mask;
+typedef FixedArray<unsigned int, GXM, GYM> map_mask;
 
 struct item_def;
 struct coord_def;
